Move `ChainPoint` to `Cardano.Wallet.Read.Chain.Point`
